reduced memory usage & added loading progress bar
- changed the underlying type of the BlockType enum to short from int
- made it so chunks that are completely empty do not allocate their block array (removes about 1/3 of chunks on mountain terrain).

So this equates to the block data taking 1/3 of the memory as before, which is cool (only a reduction of like 17MB on 10/12 render distance tho).

However, a rough estimation for 20 chunk loading distance is a saving of around 136 MB, which is pretty good.

Next would probably to do a custom collision setup and avoid unity colliders, since we can do that with only the block position data assuming I dont want to implement custom colliders.